+++ application/translations.pot
@@ -131,7 +131,7 @@ msgid "Default"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#delete()
+#: domainapp.modules.simple.dom.impl.cliente#delete()
 msgid "Delete"
 msgstr ""
 
@@ -153,7 +153,7 @@ msgstr ""
 
 #: domainapp.application.fixture.scenarios.DomainAppDemo#downloadLayoutXml()
 #: domainapp.application.services.homepage.HomePageViewModel#downloadLayoutXml()
-#: domainapp.modules.simple.dom.impl.Cliente#downloadLayoutXml()
+#: domainapp.modules.simple.dom.impl.cliente#downloadLayoutXml()
 #: org.apache.isis.applib.AbstractViewModel#downloadLayoutXml()
 #: org.apache.isis.applib.ViewModel#downloadLayoutXml()
 #: org.apache.isis.applib.domain.DomainObjectList#downloadLayoutXml()
@@ -240,7 +240,7 @@ msgid "Home Page"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#datanucleusIdLong()
+#: domainapp.modules.simple.dom.impl.cliente#datanucleusIdLong()
 #: org.apache.isis.core.metamodel.services.jdosupport.Persistable_datanucleusIdLong#prop()
 msgid "Id"
 msgstr ""
@@ -293,8 +293,8 @@ msgid "Minimal"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#name
-#: domainapp.modules.simple.dom.impl.Cliente#updateName(java.lang.String)
+#: domainapp.modules.simple.dom.impl.cliente#name
+#: domainapp.modules.simple.dom.impl.cliente#updateName(java.lang.String)
 #: domainapp.modules.simple.dom.impl.Clientes#create(java.lang.String)
 #: domainapp.modules.simple.dom.impl.Clientes#findByName(java.lang.String)
 msgid "Name"
@@ -306,7 +306,7 @@ msgid "Normalized"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#notes
+#: domainapp.modules.simple.dom.impl.cliente#notes
 msgid "Notes"
 msgstr ""
 
@@ -329,7 +329,7 @@ msgstr ""
 
 #: domainapp.application.fixture.scenarios.DomainAppDemo#openRestApi()
 #: domainapp.application.services.homepage.HomePageViewModel#openRestApi()
-#: domainapp.modules.simple.dom.impl.Cliente#openRestApi()
+#: domainapp.modules.simple.dom.impl.cliente#openRestApi()
 #: org.apache.isis.applib.AbstractViewModel#openRestApi()
 #: org.apache.isis.applib.ViewModel#openRestApi()
 #: org.apache.isis.applib.domain.DomainObjectList#openRestApi()
@@ -422,7 +422,7 @@ msgstr ""
 
 #: domainapp.application.fixture.scenarios.DomainAppDemo#rebuildMetamodel()
 #: domainapp.application.services.homepage.HomePageViewModel#rebuildMetamodel()
-#: domainapp.modules.simple.dom.impl.Cliente#rebuildMetamodel()
+#: domainapp.modules.simple.dom.impl.cliente#rebuildMetamodel()
 #: org.apache.isis.applib.AbstractViewModel#rebuildMetamodel()
 #: org.apache.isis.applib.ViewModel#rebuildMetamodel()
 #: org.apache.isis.applib.domain.DomainObjectList#rebuildMetamodel()
@@ -515,7 +515,7 @@ msgid "Type"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#updateName()
+#: domainapp.modules.simple.dom.impl.cliente#updateName()
 msgid "Update Name"
 msgstr ""
 
@@ -530,8 +530,8 @@ msgid "Value"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#datanucleusVersionLong()
-#: domainapp.modules.simple.dom.impl.Cliente#datanucleusVersionTimestamp()
+#: domainapp.modules.simple.dom.impl.cliente#datanucleusVersionLong()
+#: domainapp.modules.simple.dom.impl.cliente#datanucleusVersionTimestamp()
 #: org.apache.isis.core.metamodel.services.jdosupport.Persistable_datanucleusVersionLong#prop()
 #: org.apache.isis.core.metamodel.services.jdosupport.Persistable_datanucleusVersionTimestamp#prop()
 msgid "Version"
