+++ module-simple/translations.pot
@@ -131,7 +131,7 @@ msgid "Default"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#delete()
+#: domainapp.modules.simple.dom.impl.cliente#delete()
 msgid "Delete"
 msgstr ""
 
@@ -146,12 +146,12 @@ msgid "Domain Change"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#downloadJdoMetadata()
+#: domainapp.modules.simple.dom.impl.cliente#downloadJdoMetadata()
 msgid "Download Jdo Metadata"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#downloadLayoutXml()
+#: domainapp.modules.simple.dom.impl.cliente#downloadLayoutXml()
 #: org.apache.isis.applib.AbstractViewModel#downloadLayoutXml()
 #: org.apache.isis.applib.ViewModel#downloadLayoutXml()
 #: org.apache.isis.applib.domain.DomainObjectList#downloadLayoutXml()
@@ -199,7 +199,7 @@ msgid "Element Object Type"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#updateName(java.lang.String)
+#: domainapp.modules.simple.dom.impl.cliente#updateName(java.lang.String)
 msgid "Exclamation mark is not allowed"
 msgstr ""
 
@@ -238,7 +238,7 @@ msgid "HSQL DB Manager"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#datanucleusIdLong()
+#: domainapp.modules.simple.dom.impl.cliente#datanucleusIdLong()
 #: org.apache.isis.core.metamodel.services.jdosupport.Persistable_datanucleusIdLong#prop()
 msgid "Id"
 msgstr ""
@@ -291,8 +291,8 @@ msgid "Minimal"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#name
-#: domainapp.modules.simple.dom.impl.Cliente#updateName(java.lang.String)
+#: domainapp.modules.simple.dom.impl.cliente#name
+#: domainapp.modules.simple.dom.impl.cliente#updateName(java.lang.String)
 #: domainapp.modules.simple.dom.impl.Clientes#create(java.lang.String)
 #: domainapp.modules.simple.dom.impl.Clientes#findByName(java.lang.String)
 msgid "Name"
@@ -304,7 +304,7 @@ msgid "Normalized"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#notes
+#: domainapp.modules.simple.dom.impl.cliente#notes
 msgid "Notes"
 msgstr ""
 
@@ -324,7 +324,7 @@ msgid "Open"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#openRestApi()
+#: domainapp.modules.simple.dom.impl.cliente#openRestApi()
 #: org.apache.isis.applib.AbstractViewModel#openRestApi()
 #: org.apache.isis.applib.ViewModel#openRestApi()
 #: org.apache.isis.applib.domain.DomainObjectList#openRestApi()
@@ -414,7 +414,7 @@ msgid "Published Interaction"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#rebuildMetamodel()
+#: domainapp.modules.simple.dom.impl.cliente#rebuildMetamodel()
 #: org.apache.isis.applib.AbstractViewModel#rebuildMetamodel()
 #: org.apache.isis.applib.ViewModel#rebuildMetamodel()
 #: org.apache.isis.applib.domain.DomainObjectList#rebuildMetamodel()
@@ -506,7 +506,7 @@ msgid "Type"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#updateName()
+#: domainapp.modules.simple.dom.impl.cliente#updateName()
 msgid "Update Name"
 msgstr ""
 
@@ -521,8 +521,8 @@ msgid "Value"
 msgstr ""
 
 
-#: domainapp.modules.simple.dom.impl.Cliente#datanucleusVersionLong()
-#: domainapp.modules.simple.dom.impl.Cliente#datanucleusVersionTimestamp()
+#: domainapp.modules.simple.dom.impl.cliente#datanucleusVersionLong()
+#: domainapp.modules.simple.dom.impl.cliente#datanucleusVersionTimestamp()
 #: org.apache.isis.core.metamodel.services.jdosupport.Persistable_datanucleusVersionLong#prop()
 #: org.apache.isis.core.metamodel.services.jdosupport.Persistable_datanucleusVersionTimestamp#prop()
 msgid "Version"
